Adding ability to delete uploaded images

+++ src/com/ikai/photosharing/server/UserImageServiceImpl.java
@@ -4,6 +4,9 @@ import java.util.List;
 
 import com.google.appengine.api.blobstore.BlobstoreService;
 import com.google.appengine.api.blobstore.BlobstoreServiceFactory;
+import com.google.appengine.api.users.User;
+import com.google.appengine.api.users.UserService;
+import com.google.appengine.api.users.UserServiceFactory;
 import com.google.gwt.user.server.rpc.RemoteServiceServlet;
 import com.ikai.photosharing.client.services.UserImageService;
 import com.ikai.photosharing.shared.UploadedImage;
@@ -35,8 +38,13 @@ public class UserImageServiceImpl extends RemoteServiceServlet implements
 
 	@Override
 	public void deleteImage(String key) {
+		UserService userService = UserServiceFactory.getUserService();
+		User user = userService.getCurrentUser();
 		UploadedImageDao dao = new UploadedImageDao();
-		dao.delete(key);
+		UploadedImage image = dao.get(key);
+		if(image.getOwnerId() == user.getUserId()) {
+			dao.delete(key);
+		}
 	}
 
 }
